// src/app/store/services/pokemon/models/PokemonSpecies.ts
export type PokemonSpecies = {
  id: number;
  name: string;
  order: number;
  gender_rate: number;
  capture_rate: number;
  base_happiness: number;
  is_baby: boolean;
  is_legendary: boolean;
  is_mythical: boolean;
  hatch_counter: number;
  has_gender_differences: boolean;
  forms_switchable: boolean;
  color: {
    name: string;
  };
  flavor_text_entries: {
    flavor_text: string;
  }[];
};